Quick fixing

+++ src/main/java/com/github/mittyrobotics/conveyor2/Conveyor2Constants.java
@@ -1,5 +1,5 @@
 package com.github.mittyrobotics.conveyor2;
 
 public class Conveyor2Constants {
-    public static int TICKS_PER_CYCLE = 52900;
+    public static double TICKS_PER_CYCLE = 52900;
 }
